+++ core/src/animation/stagger.rs
@@ -227,7 +227,7 @@ mod tests {
 
   widget_layout_test!(
     stagger_run_and_stop,
-    fn_widget! {
+    WidgetTester::new(fn_widget! {
       let stagger = Stagger::new(Duration::from_millis(100), transitions::EASE_IN.of(ctx!()));
       let mut mock_box = @MockBox { size: Size::new(100., 100.) };
       let opacity = mock_box
@@ -253,9 +253,8 @@ mod tests {
       assert!(!stagger.is_running());
 
       mock_box
-    },
-    width == 100.,
-    height == 100.,
+    }),
+    LayoutCase::default().with_size(Size::new(100., 100.))
   );
 
   #[test]